refactor: move getResourceFilesize and isResourceCorsEnabled to functions/resources.ts
